Added proposed features from lsp 3.16 (#230)
* Switched from longs to ints when handling document lines and columns
* Added proposed features from lsp 3.16
* Updated capabilities to include the ability to inspect handler descriptors
* Added IStringEnumLike to deal with strongly typed strings that behave like enums but can be extended with other values based on clients
* Added DocumentUri class to replace usage of Uri.  It supports conversion from file system like paths (windows and unix), uri strings (scheme://...) and Uri.  In addition it manages the funny serialization that VSCode does in regards to : on windows file systems.

+++ src/Dap.Protocol/Models/DisassembledInstruction.cs
@@ -35,21 +35,21 @@ namespace OmniSharp.Extensions.DebugAdapter.Protocol.Models
         /// <summary>
         /// The line within the source location that corresponds to this instruction, if any.
         /// </summary>
-        [Optional] public long? Line { get; set; }
+        [Optional] public int? Line { get; set; }
 
         /// <summary>
         /// The column within the line that corresponds to this instruction, if any.
         /// </summary>
-        [Optional] public long? Column { get; set; }
+        [Optional] public int? Column { get; set; }
 
         /// <summary>
         /// The end line of the range that corresponds to this instruction, if any.
         /// </summary>
-        [Optional] public long? EndLine { get; set; }
+        [Optional] public int? EndLine { get; set; }
 
         /// <summary>
         /// The end column of the range that corresponds to this instruction, if any.
         /// </summary>
-        [Optional] public long? EndColumn { get; set; }
+        [Optional] public int? EndColumn { get; set; }
     }
 }
